conditionally implementing 'TestableContract' inside corresponding crates
