+++ contracts/mixnet/Cargo.toml
@@ -36,11 +36,14 @@ serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.23" }
 time = { version = "0.3", features = ["macros"] }
 
+cosmwasm-contract-testing = { path = "../../common/cosmwasm-smart-contracts/testing", optional = true }
+
 [dev-dependencies]
 cosmwasm-schema = "1.0.0"
 rand_chacha = "0.2"
 #rand = "0.7"
 nym-crypto = { path = "../../common/crypto", features = ["asymmetric", "rand"] }
+
 
 [build-dependencies]
 vergen = { version = "5", default-features = false, features = ["build", "git", "rustc"] }
@@ -48,3 +51,4 @@ vergen = { version = "5", default-features = false, features = ["build", "git", 
 [features]
 default = []
 contract-testing = ["mixnet-contract-common/contract-testing"]
+testing_mocks = ["cosmwasm-contract-testing", "contract-testing"]
