+++ contracts/mixnet/src/testing/mod.rs
@@ -2,3 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) mod transactions;
+
+#[cfg(feature = "testing_mocks")]
+pub mod mock_helpers;
