+++ contracts/vesting/Cargo.toml
@@ -31,11 +31,11 @@ schemars = "0.8"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0" }
 
-[dev-dependencies]
-rand_chacha = "0.3.1"
-base64 = "0.21.0"
-hex = "0.4.3"
-serde_json = "1.0.66"
+cosmwasm-contract-testing = { path = "../../common/cosmwasm-smart-contracts/testing", optional = true }
 
 [build-dependencies]
 vergen = { version = "5", default-features = false, features = ["build", "git", "rustc"] }
+
+[features]
+default = []
+testing_mocks = ["cosmwasm-contract-testing"]
