+++ contracts/vesting/src/support/mod.rs
@@ -1,1 +1,7 @@
+// Copyright 2022-2023 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
 pub mod tests;
+
+#[cfg(feature = "testing_mocks")]
+pub mod mock_helpers;
